+++ src/main/java/net/notcoded/wayfix/WayFix.java
@@ -4,6 +4,7 @@ import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.GsonConfigSerializer;
 import net.fabricmc.api.ClientModInitializer;
 import net.notcoded.wayfix.config.ModConfig;
+import net.notcoded.wayfix.util.WindowHelper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.lwjgl.glfw.GLFW;
@@ -16,6 +17,14 @@ public class WayFix implements ClientModInitializer {
     public void onInitializeClient() {
         AutoConfig.register(ModConfig.class, GsonConfigSerializer::new);
         WayFix.config = AutoConfig.getConfigHolder(ModConfig.class).getConfig();
+
+        WindowHelper.checkIfCanUseWindowHelper();
+        if(WindowHelper.canUseWindowHelper) {
+            WayFix.config.fullscreen.monitorName = "";
+            WayFix.config.fullscreen.useMonitorName = false;
+            WayFix.config.fullscreen.monitorSelector.monitors.clear();
+            WayFix.config.fullscreen.monitorSelector.monitors.add(new ModConfig.Monitors("You do not need to use this as it is automatically fixed.", 0)); // most users probably won't understand this, but I'll include it anyway
+        }
     }
 
     public static boolean isWayland() {