+++ src/main/java/net/notcoded/wayfix/mixin/WindowMixin.java
@@ -8,8 +8,11 @@ import net.minecraft.client.util.Window;
 import net.notcoded.wayfix.WayFix;
 import net.notcoded.wayfix.config.ModConfig;
 import net.notcoded.wayfix.util.DesktopFileInjector;
+import net.notcoded.wayfix.util.WindowHelper;
 import org.lwjgl.glfw.GLFW;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -39,21 +42,24 @@ import net.minecraft.resource.InputSupplier;
 */
 
 @Mixin(Window.class)
-public class WindowMixin {
+public abstract class WindowMixin {
+    @Shadow protected abstract void onWindowPosChanged(long window, int x, int y);
+
+    @Shadow @Final private long handle;
+
     @Inject(method = "<init>", at = @At(value = "INVOKE", target = "Lorg/lwjgl/glfw/GLFW;glfwDefaultWindowHints()V", shift = At.Shift.AFTER, remap = false))
     private void onWindowHints(WindowEventHandler windowEventHandler, MonitorTracker monitorTracker, WindowSettings windowSettings, String string, String string2, CallbackInfo ci) {
         if (isWayland()) {
             GLFW.glfwWindowHint(GLFW.GLFW_FOCUS_ON_SHOW, GLFW.GLFW_FALSE);
 
-            if(!WayFix.config.injectIcon) return; // assuming that user wants wayland icon and not broken icon (mc launcher)
-            DesktopFileInjector.inject();
+            if(WayFix.config.injectIcon) DesktopFileInjector.inject();
             GLFW.glfwWindowHintString(GLFW.GLFW_WAYLAND_APP_ID, DesktopFileInjector.APP_ID);
         }
     }
 
     @Redirect(method = "updateWindowRegion", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/util/MonitorTracker;getMonitor(Lnet/minecraft/client/util/Window;)Lnet/minecraft/client/util/Monitor;"))
     private Monitor fixWrongMonitor(MonitorTracker instance, Window window) {
-        return WayFix.config.fullscreen.useMonitorName ? getMonitor(instance) : instance.getMonitor(window);
+        return WayFix.config.fullscreen.useMonitorName && !WindowHelper.canUseWindowHelper ? getMonitor(instance) : instance.getMonitor(window);
     }
 
     @Unique
@@ -68,6 +74,18 @@ public class WindowMixin {
         }
 
         return instance.getMonitor(monitorID);
+    }
+
+
+    // KDE Plasma ONLY
+    @Inject(method = "updateWindowRegion", at = @At("HEAD"))
+    private void fixWrongMonitor(CallbackInfo ci) {
+        if(!WindowHelper.canUseWindowHelper) return;
+
+        int[] pos = WindowHelper.getWindowPos();
+        if(pos == null) return;
+
+        onWindowPosChanged(this.handle, pos[0], pos[1]);
     }
     
     @Inject(method = "setIcon", at = @At("HEAD"), cancellable = true)